FastAPI开发/FastAPI_Test/requirements.txt
fastapi
uvicorn
python-socketio[asgi]
PyQt5
